Added purge_stale_applications command
When a URL is changed on an application, the check-link
record for the old URL remains in the check_link_report
table.

This function is used to remove check-link records for
application links that have not been checked since
'older_than'.  This function is called immediately after
doing a check-applications with older_than set to the
datetime when the check-applications process began.

It can also be invoked from the CLI with an arbitraty older_than value.